# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API сервер для управления задачами",
        "title": "ToDo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/newtask": {
            "post": {
                "description": "Создать новую задачу с указанными данными",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создать новую задачу",
                "parameters": [
                    {
                        "description": "Данные задачи",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "description": "Получить список задач с пагинацией и фильтрацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить список задач",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество элементов на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Статус задачи (true - выполнена, false - не выполнена)",
                        "name": "completed",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата в формате YYYY-MM-DD",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TasksList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Получить задачу по её идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить задачу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновить существующую задачу по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Обновить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные задачи для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить задачу по её идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.Response": {
            "description": "Ответ обработчика",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные ответа"
                },
                "error": {
                    "description": "Сообщение об ошибке (если есть)",
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "description": "Статус ответа (OK/Error)",
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "models.Task": {
            "description": "Задача пользователя",
            "type": "object",
            "required": [
                "due_date",
                "title"
            ],
            "properties": {
                "created_at": {
                    "description": "Дата создания",
                    "type": "string",
                    "example": "2025-04-17T10:30:00Z"
                },
                "description": {
                    "description": "Описание задачи",
                    "type": "string",
                    "example": "Купить 2 литра молока в магазине"
                },
                "due_date": {
                    "description": "Дата выполнения",
                    "type": "string",
                    "example": "2025-04-20T15:00:00Z"
                },
                "id": {
                    "description": "Уникальный идентификатор задачи",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "description": "Статус выполнения (true - выполнена, false - не выполнена)",
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "description": "Заголовок задачи",
                    "type": "string",
                    "example": "Купить молоко"
                },
                "updated_at": {
                    "description": "Дата обновления",
                    "type": "string",
                    "example": "2025-04-17T10:30:00Z"
                }
            }
        },
        "models.TasksList": {
            "description": "Список задач с пагинацией",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Массив задач",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Task"
                    }
                },
                "limit": {
                    "description": "Количество элементов на странице",
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "description": "Текущая страница",
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "description": "Общее количество задач",
                    "type": "integer",
                    "example": 42
                }
            }
        },
        "response.Response": {
            "description": "Стандартный ответ API",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке (если есть)",
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "description": "Статус ответа (OK/Error)",
                    "type": "string",
                    "example": "OK"
                }
            }
        }
    }
}